Abstracted kernel version information.
Also, fail on errors, except for handled, known-bad commands.

Why're we allowing failures in some commands?  They'll never work,
because they depend on the actual DreamPlug hardware, but we can't
boot the built image on the actual hardware unless these commands run.
So, cheat: run them as far as they'll go, because that's just far
enough.