Prep updates $NAME, all archs use simple-image.
So I got tired of the makefile repeating itself and folded each
architecture into prep's logic, while all the images now depend on the
simple-image target.  Also, the comments are now a little more
informative.

The biggest problem is that we can't run multiple targets at once any
more, they need to be in different processes, because prep doesn't
keep separate state for each process: it configures the environment
based on whatever the current goal is.